Add parsing of [Headers] section
Adding parsing of the next section in
a file - the [Headers] section.

Uses captureSection with the new
trim parameter (since headers should be
trimmed - only in the body of a requst
should the whitespace be significant).

Once lines are captured, check for duplicates
but only have them as a warnings for now.

Test plan:
* go test -count=1 -v github.com/jonaslu/ain/...
  passes.
* go run main.go test.ain prints out the Headers
  entered in test.ain. Verify that the last line
  is used.