Start afresh - add reading of the template file
This is the same functionality as in comit
1c3b7b0c, only with the added check that
if we're not connected to a pipe the main
program prints an error if no file-name
has been given.

And the addition that the given file name
over a pipe or as a command line argument
is now trimmed for any whitespace (as we
cannot find and open the file if it contains
newline or spaces).

Test plan:
* go run cmd/ain/main.go prints an error-message
  on missing template-file name.
* echo 'test.ain' | go run cmd/ain/main.go opens the
  file in the $EDITOR and then prints the contents
  of the edited file.
* echo '    test.ain ' | go run cmd/ain/main.go
  works as the test-case above.
* find . | fzf | go run cmd/ain/main.go does the
  same - opens the given file in the $EDITOR
  and prints the contents.