Add back parsing of the template + fatals
This is the same contents as in commit 67c1431
with some additions:

Since errors is a keyword in go I want some
way to distinguish between true go-errors
(such as a file cannot be opened) and user errors
because the template is malformed or doesn't contain
the right values.

These are called fatals - a fatal is a user error
that stems from a malformed template. With the template
being the file that the user edits before any backend-call
is made. Backends are curl or http:ie, it's what the
data in the template will be transformed into - a call
to a backend resulting in some html or json or
what have you.

The main.go program uses disk.ReadTempalate to get the
raw string from the user. The parse.ParseTempalte then
does all the lifting when transforming the template into
the call.Data structure needed by the back-end. This means
trimming any whitespace, parsing and validating the known
headers and prints returns any fatals if encountered
so the main program can print them and exit nicely (status
0 since it's a user error).

Unit tests-proves that the [Host] and [Headers] sections
are still what they used to be in commit 67c1431 but with
an important exception - warnings have been dropped.

We now exclusively deal with fatals and warnings are gone.
Either the template passes in it's totality, or we exit directly.

Fatals are formatted as it would be in a compiled program, with
line number - a marker on the errouneous line and the error-message
first, with the exception if there a source line index would not be
useful (such as a section appearing repeatedly). When this happens
only the error-message is displayed.

Test plan:
* go test github.com/jonaslu/ain/... passes.
* go run cmd/ain/main.go test.ain with an empty [Host]
  section prints a fatal message.
* go run cmd/ain/main.go with an empty [Host] and two
  [Headers] sections prints two fatals.