Add error on empty template
Make ain handle empty templates
(give an error). Also - if the
last line in the template contains
a newline this will make error-messages
print an empty last line if the error
occurs at the last or last-1 line.

Remove it so error-message does not
display an erroneous last line.

Test plan:
* Run ain with an emoty template
  (zero and serveral empty lines).
  Verify an error-message.
* Have an error (such as an empty [Host]
  header first, middle and last in the
  template. Verify it prints line-numbers
  and no erroneous last line.