Add transform stage before parsing
The transform stage looks for environment
variables and inserts them into the template
before it's parsed. This gives opportunities
to insert variables at runtime (e g
ENV=test ain endpoint.ain).

Anywhere in the template where the symbol ${ENV}
is encountered it's replaced with the test value.

What the format of a shell-variable is seems
to be variying so I'll be optimistic and
allow anything and try it (even with =)
instead of having an opinion if it's allowed
or not.

https://stackoverflow.com/questions/2821043/allowed-characters-in-linux-environment-variable-names

Test plan:
* Verify empty variable ${} gives an error-message.
* Verify not set variable ${NOTEXISTS} gives an
  error-message.
* Verify EMPTY="" and use ${EMPTY} in the template
  gives an error-message.
* Verify a line with one valid ${EDITOR} ${BROKEN}
  gives an error-message.
* A variable spannig several lines:
  ${VAR
  } gives an error-message.
* An existing variable (such as ${EDITOR} on the
  same line twice works.