Add shell command capture to template
Add a $() functionality to the template where
anything inside the $() is run as a command
and it's stdout output is captured and inserted
into the template.

Find all $() commands and capture their
command and possible arguments. If
something is malformed such as a missing
command, make it a fatal and return to the main
program.

Then, run all commands in parallel with some
timeout set (context passed in will also be used
to trap any SIGINT etc and cancel any outstanding
shell commands). If any command fails, make it
a fatal by capturing both it's stdou and it's
stderr. Insert the result into an array that
will have the same ordering as it was captured.

This way two commands on the same line will be
correct in order.

Then insert the result of the commands stdout
into the template.

Test plan:
* Run ain with an empty shell command:
  $(). Verify error message and program exits.
* Run ain with some botched shell command
  (e g  $(MULTILINE
  VARIABLE)). Verify error message and program exits.
* Run some program with only no arguments (e g ls)
  and verify its stdout is captured and inserted into the
  template.
* Run program with something that exits with
  an error-code (such as 'bash -c 'exit 1'').
  Verify that the error from the program
* Run a program that produces no stdout output.
  Verify ain exits with an error-message on
  no result captured and exits.
* Run ain with some more complex command that requires
  a shell as done above (e g bash -c "sleep 1 && echo yak"). Verify
  that program captures output and can interact with the
  shell.
* Do the same as above but with single-quotes.
* Run or write a small program that takes the same arguments
  but produces different results on each invocation. Verify
  that the command output is inserted in the correct order
  in the template.