Add curl as the first backend
Our first backend curl, takes the call.Data
and massages it into a curl call. Currently
only the URL and headers are useful.

Then capture the output, it it fails capture
both stdout and stderr and return then as an
error and if it succeeds only return stdout.

Have a timeout for the backend (curl supports timeouts to,
but to be on the safe side (such as curl pointing
to another binary) have a timeout in ain too).
Also to hook into the SIGINT / SIGHUP etc
we need a context with a timeout anyways.

Test plan:
* Make curl error out (invalid url for example).
  Verify ain exits with status 1 and the error-message
  from curl is printed to the terminal.
* Make curl not error out and verify that the
  output is printed to the process stdout.
* Use ain in a pipe (to jq for example) and verify
  result is pipe:able forward.