Add httpie as a backend
Use template strings for building the
backend command. Not sure about this,
it seems brittle, but it will
allow me to add more backend-commands
with ease.

Same procedure as with the shells,
split the string on space, add back
any quoted strings and run it.

Creating the utils-package for this
that handles unsplitting of strings.

Test plan:
* Run with a non-existing backend.
  Verify outputs error message and
  exits.
* Run a backend-command that exits
  with some error (such as a non-existing
  host). Verify both stdout and stderr
  is captured and program exist with error.
* Run working command (with headers)
  for both curl and httpie and verify
  result is returned successfully.