Properly tokenize strings
Strings with quotes in them
are of course a bit more tricky
than what I naively assumed.

Build a proper string scanner that
removes space and keeps quoted
(both single and duoble) quoted
strings intact (akin to what bash does).

This is done by scanning the incoming
string, if it's space igore it and move
on. If it's not check if it's a special
quote character.

If so save what character we are looking
for to end the quoting.

Continue parsing the string. If we've saved
off a quote character, check if it's a balancing
such at the current position and that the prevoius
character wasn't an escape cacharacter (\).

If not, see if we're at space and capture the
word.

At then end, if we're still looking for a terminating
quote character then that's an error.

If not collect the last word.

Test plan:
* Unit tests for all thinkable cases.