Add optional removing of quotes
This is what bash does, it runs
commands and it needs to know what
strings it should keep intact as
one string and what strings that
it shuold split and pass as separate
arguments.

So in the backend template quotes are
used to keep an argument intact, but
the quotes themselves needs to be removed
when strings are extracted so the content
of the string is passed to the command
and not the quotes themselves.

Add a boolean to indicate wheter the
quotes (and any whitespace inside the
captured string should be removed).

Test plan:
* Some more unit tests for removing
  quotes inside the string.