Adding [Body] section
Body is optional and is kept verbatim
(not trimmed) since whitespace
may be significant.

As whitespace may be
significant, write it to
a temp-file and send it as file-data
(mangling it into a single-line command
is afaik not doable).

Not that testing failing to write
to a file that has been already
opened is not possible without severe trickery,
so I'll assume it works:
https://unix.stackexchange.com/a/351625/50004

Test plan:
* Two body sections, verify ain
  prints an error-message and exits.
* Have some body-content and hack the
  tempfile-creation and set it to
  a non-writeable folder. Verify ain errors
  out with "Could not create tempfile" and
  exits.
* Omit body and verify no temp-file is
  created or included in the backend-call
  by commenting out the tmpFile.Close()
  (file should be left in the temp-directory
  if used in the call).
* Verify the happy path (such as having
  some valid json with the header
  Content-Type: application/json) and
  POST it to an endpoint. Verify both
  curl and httpie can successfully post
  content.
* Run the call above and verify temp-file
  is removed when ain exits.