Adding [Backend] section
Single value section on what
backend to use when making the
http-call.

This section is mandatory.

Store the function to call
in the callData struct. Not
too happy about this, it has
a weird redundancy feel to
it (structure holding a reference
to what it's being passed to) but
can't think of a prettier solution
for now.

Test plan:
* Verify No [Backend] section
  gives an error and ain exits.
* Several lines under backend
  gives an error and ain exits.
* Verify curl uses curl as
  backend.
* Verify httpie uses http(ie)
  as the backend.
* Verify unknown backend
  gives a fatal on unknown
  backend.