Refactor: make backends into an interface
The reason for the refactor
is that I want the backend-timeout
to be handle uniformly
(call the backend, check the
deadline exceeded error before
checking the backend error).

Refactor of the call,
have an interface for
the backends and a
constructor for each.

In the parsing I need to check
if the backend is valid. Don't
be cute, use a method located
close to the constructors.

Then, get the backend
which is now an interface -
return an error if the
backend is not known
(should not be possible
to get here unless the
parsing and validation of the
backend name and the
instantiation have diverged).

Let the backend assemble the
command and run it with the
timeout.

New is that the backends
now leave the tmp-file
and does not remove
it when done. Will
be fixed in an upcoming
commit.

Test plan:
* Invalid backend name
  in a template results
  in error.
* Smoke test for regression
  on curl and httipe
  templates.
* Cause a timeout when running
  a backend (such as curl
  --limit-rate 1). Verify
  timeout-message now includes
  the backend name.