Refactor:
In preparation of adding a
global file, refactor the
reading of the local template
(as opposed to the global template)
so that read returns the file-name
and not the read template.

This is so we can parse
multiple templates (local
and global).

Also:
Check arguments first, so even if
we're connected to a pipe any given
file name trumps what's been sent
down the pipe.

Test plan;
* Verify that no local template
  file name or pipe gives an
  error.
* Verify that file-name and
  connected to a pipe favours
  the file-name over the
  name sent across the pipe.
* Verify a name sent over
  a pipe is accepted if no
  filename is given.