Make [Host] section append
Very often different templates
will have the same base-url,
and some minor variation
on the path.

By making host append, we
loose the ability to pinpoint
the template in which
where the malformed URL was
given but gain that a
service base url can
be put in a base.ain file
and the sub-path appended
in the variant files.

E g ain base.ain post-update.ain
where base.ain contains
[Host]
http://localhost:{PORT}/

and post-update.ain
contains
[Host]
/api/new-member

Test plan:
* Have a template with an
  Empty [Host] section. Verify
  ain exist with an error-message
  on empty [Host] section.
* Have a single template file
  with an invalid url (e g bllkj aldkjf).
  Verify ain exits with an error-message
  on malformed url.
* Do the same as above but spread
  out the invalid url over several
  files.
* Have a single template file
  with a well-formed url. Verify
  ain is successful.
* Have a template with
  the well-formed url
  spread out over several files.
  Verify ain is successful.
* Smoke test config marge.