Ooops: validate merged sections unless errors
With merged sections across
several files we now have two
stages of errors, one is
errors in the individual files,
and one is errors in the
aggregated sections.

Only validate the aggregated
sections when there are no errors
in the individual files (merge-data
may be nil - e g if the template
is empty).

Rather than trying to merge
and validate something incomplete
only do that when the individual
files are ok.

This means however that with only
one file (no merges) any errors
in the "merged" sections will
only be presented when the other
errors are fixed (any errors
in the section parsers). This is
acceptable for now and can be
fixed if need be.

This also marks a design-decision
that whenever there's an error
returned the other return
values of a function are void
(regardless of wether it's null
or not).

Test plan:
* Have an error that returns
  nil as the parseData (such
  as an empty template).
  Verify no segfault as
  no merge takes place
  unless there are no errors
  in the individual files.