Drop -x flag, use file! for one-shot editing
Since ain now supports multiple templates
-x becomes an all or nothing option.
We cannot edit one or some of the
files.

Replace it with a special character
at the end of the file-name (!) to
mark that this file should be edited.

Neato-trick: if we place it at
the end of the file-name we
can use bash:s default auto-complete
on file-names and simply back-up
to mark it as open for editing.

Test plan:
* Run ain with ain template/file!.
  Verify that the contents of that file
  is opened in the $EDITOR.
* Run ain with template/file template/file2!.
  Verify that the contents of the second
  file is opened in the editor.
* Run ain with ain template/file. Verify
  no $EDITOR is opened.