Prepare curl for print as command
Convert so backends can either execute or
print itself as a string (so
it can be passed to your BuddyPal).

To do this we want the arguments
to the backend escaped when printing
as a string (special chars, injection etc).

But in order to DRY, move out the
construction of the arguments to
separate methods that can either be
escape or not.

Prepare for the print-out by
making the headers and backedoptions
[][]string (so we can pretty-print one
entry on each line).

Part of the change is also to spit
out the body-file in the same
directory as the printed command
so it can be IM:ed wholesale, and
also work as a cheap-o debug.

Test plan:
* Pass a template that has backend-options
  method, headers, body and curl as a backend.
  Verify that all are still passed to CURL by
  replacing curl on the path with a command
  that dumps out all the passed variables:
  curl.sh
  echo $@
  export PATH=<local-curl-path>:"$[PATH}
* Do the same with httpie and omit --ignore-stdin
  and verify any backend-options are still passed
  correctly.
* Add --ignore-stdin to httpie template and
  verify it's not added twice.