Add a util method for escaping shell strings
In order to print the commands correctly
to the terminal, any special characters
must be escaped (&,? etc). Go the
easy route and more future-proof
route and quote with single quotes
regardless of contents of the string.

The only thing needed to be escaped is
the single quote which is done by
the '"'"' hack.

Test plan:
* Have a string with a quote
  in. E g it's fine. Implement
  a test and print the escaped
  content. Paste it into
  an echo <string>. Test it
  in bash, sh and zsh and
  verify that the string prints
  correctly (the shell does not
  intercept any special caracters).