Bug: always trim quotes from subshells
We need to be doing exactly what a
shell is doing with the passed parameters
in a subshell. The first command is what's
run (as we're using command.Exec), the rest
are arguments.

In order to have an argument to a command
contain spaces we need quotes, but those
quotes needs removal before passed to
the command.

Test plan:
* Insert $(bash -c 'echo "{\"test\": 1} | jq ".test"')
  into a template. Send it as a header for example.
  Verify the subshell now completes and returns
  1.