Conclude line-endings conundrum
What to do with line-endings?
Add a note in the README and
drop the onus on the user.

Because it's nigh impossible to
guess the encoding on based on OS,
and most programs on windows
have accepted the fact that there
are other OS:es out there, let
it be for now.

Test plan:
* N/A