Add regex to test if wget outputs to stdout
Same story as with httpie, add the -O- backend option
which directs wgets output to stdout but
only if the user hasn't specified it. It is valid
to add it twice, but when sharing with the -p
flag it might confuse if there's two of the -O
options.

Test plan:
* Have a template that uses wget as a backend.
  Run ain -p <template.name> and verify that the
  -O- option is added first to the list.
* Have a template with the options -O-, -O -
  respectively in two templates. Verify with the -p
  flag that -O- is not prepended in this case.
* Have two templates with some extra options
  into the same - flag, E g -qO-, -qO -.
  Verify that no extra -O- is added when running
  the -p flag.