Move binary name outside backends
The binary name is repeated in the backends
(once for calling it and another for when
printing as a string). Move it to outside
the backend itself.

This allows us to iterate over the map
with the template names and get only
the binary name for testing if the
binary is present on the box or not.

Test plan:
* Run ain on each backend and verify
  that the call is still valid.
* Run ain -p on each backend, eyeball
  the output and verify the binary is
  correct and that there is a space
  between the binary name and the
  rest of the command line.
* Run ain -p for each backend and
  pipe it into bash again. Verify
  that it produces the same output
  as minus the -p and | bash.