Add a [Query] section
Because some query-parameters might
be shared in several paths - add
them as their own [Query] section.

An example of why this is useful is
when an API_KEY needs to be sent to
several endpoints.

The [Query] section is left and right
trimmed and added to the url.RawQuery.
It's appended with an ampersand if
there already are some query-string
and no ampersand if it's the first
query-string.

Test plan:
* Have an URL with no query parameters
  and add a [Query] section to the template.
  Add one query parameter.
  Verify via ain -p template.ain that the
  one query parameter is appended to the URL.
* Add another query parameter to the [Query]
  section above. Verify via ain -p that
  there are now two query parameters separated
  with an ampersand.
* Have an URL that ends in a ? (forceQuery)
  and at least one row under the [Query]
  section. Verify via ain -p that the
  query parameters are added to the URL.
* Add another [Query] section in a new
  file and run ain -p <first-file.ain>
  <second-file.ain> and verify that
  the query parameters combine over
  files (in the order of appearance).
* Do the same as above but have an
  URL that already have some query
  string in it.