Make query-string delimiter configurable
Add a config setting for what delimiter
to use for query-strings. It defaults
to & (the by far most common use case),
but can if wanted be set to ; or any
other sequence and even then empty
string.

It goes under the [Config] section
and has the format
queryDelim=<string with no spaces>

Spaces are disallowed in the string
(because the delimiter is not
url-escaped).

Test plan:
* Define the queryDelim=
  twice in the same file and verify
  an error message is produced.
* Have a [Query] section with at least
  two values. Define the queryDelim
  to be =  and verify that values
  are now concatenated via ain -p.
* With the example above,
  verify whitespace is ok
  anywhere between queryDelim,
  = and end of string by defining a delimiter
  and using ain -p to print out
  the result.
* Set the queryDelim= val pal and verify
  that an error message is produced that
  the query delimiter cannot contain spaces.
* Omit the queryDelim in the [Config]
  and verify that the default '&'
  is used to concatenate values.