Properly url-encode the query string
URL-encoding strings is a pain in the neck
and is something that ain will
help with as much as possible.

Ain merges all query parameters
in the [Host] and [Query] sections
into one huge string.

Then it splits it on key-value
pairs given that the delimiter is
not the empty string.

It then encodes any key-value pairs
or the entire string if the delimiter
was the empty string.

It does the encoding as follows,
scan the string and replaces
any character outside the
US ASCII-set except for %<hex><hex> and
+.

This means that even half-encoded
strings (such as Göran%2F)
will be fully escaped (
in this case G%C3%B6ran%2F).

The only caveat that the % and
+ chars themselves are ambiguous.
Did you mean %20 literally?
Or did you mean Alice+Bob
literally? In these cases, you
need to yourself encode the
percent sign (%25) and
or plus sign (%2B).

Test plan:
* Verify that queryEscape escapes
  illegal characters (such as ;).
* Verify that queryEscape does not
  escape the plus sign (+).
* Verify that queryEscape does
  escape a non-hex sequence (%, %1).
* Verify that queryEscape does not
  escape a valid hex sequence.
* Verify that queryEscape does
  not escape chars within the
  US ASCII range.
* Verify that a space is encoded
  into the + sign.
* Have a template with a [Host]
  URL that has no query-string,
  and no [Query] section. Verify
  url is still correctly assembled
  via ain -p.
* If the queryDelim = "", verify
  that the entire query-string is
  encoded by having an illegal
  character such as ; in the [Host]
  and [Query] section.
* Set the queryDelim = ";" and
  have no key-value pairs (string
  contains =) but have illegal
  characters in the [Host] and
  [Query] strings. Verify with
  ain -p that the whole string
  except the delimiter is encoded.
* Finally have an equal-sign
  and use the default delimiter
  (&) with some illegal character
  in both key and value
  in [Host] and [Query]. Verify via
  ain -p that key-values are encoded.