Trim key-value whitespace in [Query] section
Let's avoid bash-isms shall we? The whitespace
around equals-sign (=) in the query-string
key-value in the [Query] section is currently
significant.

This will be confusing, better to let the user
actually encode spaces themselves if need be,
or put that key-value in the [Host] section
as it is significant there.

Make this apply only to the [Query] and
not the [Host] section as it'll be weird
if spaces are significant in the url-path
but not the query in [Host]. A space
is a space there.

Do this by manually splitting the host.RawQuery
if the delimiter is known, encode it
with no spaces in between the key-values.

This enables us to have the empty delimiter
for the [Query] sections and still
encode the key-values so it only
applies to the [Host] section.

Test plan:
* Have a template with a regular
  url and no query part.
  Verify no regression.
* Have an URL with a query string
  ca$h=yes in the [Host]
  section. Set the queryDelimiter
  to "" and verify that the entire
  string is encoded.
* With the case above. Set the delimiter
  to != "" and verify that the key-values
  are now encoded.
* With the case above add a space
  so the query-part becomes
  ca$h = yes and verify that
  the spaces are encoded as +.
* Add the same key-value as above with
  the correct delimiter and verify
  that the key-values are encoded.
* Have an URL and the values
  ca$h=yes in the [Query]
  section. Set the queryDelimiter
  to "" and verify that the parts are
  encoded but the = is kept.
* With the case above. Set the delimiter
  to != "" and verify that the key-values
  are still encoded.
* With the case above add a space
  so the query-part becomes
  ca$h = yes and verify that
  the spaces are discarded.
* Add the same key-value as above
  and verify that the key-values are encoded
  and correctly joined with the delimiter.