Fix for incomplete quoting
Quoting was naïve and did not handle
when quoting in the middle of a word
(e g wo" yak "rd), which bash handles.

To keep quoting as bash-like as possible
re-do the quoting algo. Using a string builder
instead of slice indexes to simplify things.

Adding unit tests because they make sense
for once. This is the canonical case,
to exercise and prove an algo working.

Test plan:
* Unit tests (gasp) pass.
* To smoke test:
  Run echo some\" file \"name | ain
  and verify that the file name
  is interpreted as `some file name`.