Add did you mean on environment variables
Try to be even more helpful when user
gets it wrong. Add a levenshtein distance
calculation (homebrewed, not optimal but
I understand it) to find something that matches.

Print up to 3 variables that have the distance
2 or less if found. Ignore if the environ is
really large (might take time to find matches).

Optimize some by not comparing strings that
differ too much in length, as they will never
come under the threshold.

I thought about limiting the search
if the environ is too big, but let's wait
and see if that is a real problem.

Test plan:
* Have a template with some variable, such
  as PORT. Make sure en environment is
  clean from matching variables. Run
  ain and verify that the same error-message
  without suggestions is printed.
* Setup as above, but run PROT=3 ain template.
  Verify that ain suggests 'Did you mean PROT'
  in the errormessage with no variable found.
* Setup as above, add 2-4 similar sounding
  environment variables when the program is
  run and verify that up to three suggestions
  are added.
* As above but have a variable APORTB=2 and
  verify it's included in the error-message.
* As above but have a variable APORTBC and
  verify it's not included in the error-messsage.