Refactor: remove sectionHeaderLine from captureResult
This is a remnant from when there could be
only one .ain file and not specifying the
header was an error. This is now done when
the templates are assembled.

Also, since we're getting the sectionLines
from capture straight away, don't rename
it to backendLines etc , it's clear enough
what it is.

Test plan:
* Have some template with an empty section
  (e g [Backend]). Add an fmt.Println("yak")
  and verify that empty section lines returns
  early from the method. This verifies that
  not having the sectionHeaderLine == emptyLine
  works as before,
* Some cursory smoketesting since the
  rest of the change remove explicit
  naming of the sectionLines (i e this
  is where static typing shines).