Bugfix: allow for empty file.ain!
Strange - ain relies on the number
of bytes written, but wraps err
in the errors.Wrap.

This caused an empty file.ain!
to not open in the set $EDITOR
because the written number of bytes
were zero. But still a valid file.

According to the docs, if err != nil
then there was an error. Also the
errors.Wrap(err) will return non nil
because err is no nil.

Also: remove named return variables
as they weren't used.

Test plan:
* go run cmd/ain/main.go empty.ain!
  opens in the $EDITOR. Enter no
  text and save. Verify ain errors out
  with a "cannot process an empty template".
* Do the same but add some valid sections.
  Verify call happens.
* Do the same with an already valid file.
  Verify $EDITOR opens up and call happens.