Refactor: add fatals as a field to sections
Fatals (it's your fault) are in the first parsing
stage linked to a particular file.

Thus, making sections carry their own fatals[]
and being able to format fatals based on the
raw template makes sense.

This way fatals does not have to be a sourceMarker with
an emptyLine which is good. Fatals are just a string
to discern it from an error (which is ains fault).

Test plan:
* go run internal/pkg/sections/*.go empty.ain.
  Verify it errors out with an error-message
  "Cannot process empty template".
* The same but with a non-empty template file.
  Verify structure is dumped out (execution ok).