Refactor: add validation of headings
Move the capturing of sections out
to it's own func for better readability.

The temporary capturedSections data-structure
is merely this: what heading, what line it
was found on and the rows below it until
the next section.

Using this we can easily find out
empty headings (no sourceMarkers registered)
and we can check for duplicate headings
defined.

To not regress we're using the same
formatting output the as before
even though we now know the
lines where the headings were defined.

The nice thing which is technically
a regression is that errors in the parsing
of the templates will come in the same
order as the file and not "random"
as when parsing heading by heading.

Also: small oops, template not is "empty"
when the first header is found.

Test plan:
* Have a template with some empty headings.
  Run main and verify error message on all of these.
* Have a template with some multiple defined headings.
  Verify error messages on section defined twice.
  This is also a regression since line numbers for
  multiple sections were actually off in the original
  version.