Set contents int the sections struct
If we're ok (no fatals) set the []SourceMakers
in the sections struct. Trim all contents except
body (whitespace is kept) as per the original.

Having the capturedSections struct holding
a pointer enables us to set it directly
when hitting the section heading so
we can append to it or leave it empty if none
found. If not using a pointer we need to keep
track of when the next heading is hit or the file
ends - which makes it a bit more cumbersome.

We've got some options when ranging over a
slice of structs since the range variable
is a local copy of the struct in the slice.
Either copy the struct and return a new slice,
or manipulate the items in-place. Opting for
manipulating them in-place.

This requires using the idx of the slice
and manipulating what the slice points to.

Test plan:
* Have a valid template with indentation on
  all valid section lines.
  Verify that the section "Sections" are set
  and that all content except for the body
  is trimmed of any space.