Make Sections.sections into a map
And use constants to access that map.
A map makes more sense, adding
a new section will be as simple
as adding a string, there is only
one reference to it in the code,
and I can iterate through all or
some sections with the help of the map.

Also - convert the []SourceMarker
to a pointer as we'll be manipulating
the contents in-memory and not copying.

Test plan:
* Have a template with all valid sections.
  Run go run internal/pkg/sections/*.go test.ain
  Verify that the output contains all valid headings
  and that the sourceLineIndex points to the
  correct line (minus one) in the valid template.