Replacing of envVars in the SectionedTemplate
Range over all sections in a SectionedTemplate
and perform replacement of envVars.
The envVars replacement is an almost 1:1
copy of the original, but naturally
driven by sections instead of on the
raw template-file. And setting s.SetFatalMessage
instead of returning them.

Also: the SectionedTemplate is turning
more and more object-oriented. I don't
know if I like it yet.

Test plan:
* Compare the two envvars files. Verify
  that the only difference is sections / rawTemplate,
  setting fatals and how the result is returned / stored
  instead.
* Have a template with some env-vars, and at least
  one line with several env-vars. Run go run
  internal/pkg/sections/*.go envtemplate.ain
  without any values set. Verify errors on
  all places.
* As above but with all values set. Verify
  the resulting sectionedTemplate now
  contains the envVars substituted for their
  values.