Add back dynamic expansion
Right now if an env-var or an
executable returns a section
(e g VAR="[Body]\nBody body)"
and a template including ${VAR}
the returned [Body] is ignored.

Bring that back, but do it
in an expanded context in the
sectionedTemplate so we can also
print out any expanded context when
printing a fatal (since error-messages
not including the expanded context
are not as helpful as we can be).

Start off the change by adding in an
expandedSourceMarker - which holds the
sourceMarker and an expanded flag if
the line has any expansions (env-vars
or executables).

This will be used in an upcoming commit
when printing the new context.

Test plan:
* Have an empty template. Run
  go run cmd/ain/main.go empty.ain.
  Verify error on cannot process
  empty template.
* Have some template with at least
  two headings and some content.
  Verify via inserted printing of the
  sectionedTemplate that the expanded lines are
  the same as the rawTemplateLines, but with
  comments and whitespace lines removed.
  Verify the index in the expanded template
  lines points correctly to
  the rawTemplateLines slice.