Re-do capturing to selected sections
Remove the split-and-trim section.
This is done last in the capturing
phase - since envvars and commands
can return more whitespace and
comments.

Re-work the capturing to
not capture all sections, but
only a list of wanted sections.

This allows us to iterate over
the expandedTemplateLines several
times, lifting out sections
several times (while the expanded
content can change in-between).

Test plan:
* Have a template with two
  sections, e g [Config].
  Set s.setCapturedSections(configSection).
  Verify error on multiple headings found.
* Have a template with an empty
  [Config] section. Verify
  error message on empty heading.
* Have a template with at least
  two different sections.
  Set the s.setCapturedSection
  to capture these two sections.[
* Have a template with a [Host]
  section. Add an escaped \[Host]
  line below the section and verify
  that sections can be escaped.