Use captured executable expressions
Instead of running a regexp
on the line twice, once for
capturing and once for inserting,
save off the result on the
executableAndArgs struct and
use the captured result to
find the matching line again
when inserting. This will
work since the iteration order
is the same both times.

Copy over the executableExpression
to the executableOutput struct
so we can re-use it to find the matching
line when re-inserting.

Then, for the first item in the
list, keep searching the expanded
template lines again until we find
it first in the string.

Since we can have many executables
on one row, cut the noCommentsTemplateLine
for each match.

Then advance the executableOutput
array for each match until there
is none left and we're done inserting
back the result.

Test plan:
* Have a valid template with no
  executables. Compare this
  commits version with the previous
  commits (or a baseline ain) and
  verify the output is the same.
* Have a two templates (or
  reuse the same twice) with
  at least two errors
  in the executables
  e g $(cmd 'arg1) and
  $(cmd yo. Verify
  both errors are aggregated
  before ain exits, with the
  correct line number and expanded
  context.
* A template with at least
  two failing commands
  (e g ${exit 1} $(exit 2)).
  Run the template twice and
  verify ain aggregates the errors
  on the correct lines
* Happy case with at least
  three executables, and
  two of them on the same line.
  (e g $[printf 1] $(printf 2)
  $(printf 3)).
  Verify result is inserted in the correct
  order.